fixed critical bug that made ransac never pass any model other than perfect
blindly removing float casts lead to integer division which always
returned 0 except for when all matches were correct

+++ mpicbg/src/main/java/mpicbg/models/AbstractModel.java
@@ -128,7 +128,7 @@ public abstract class AbstractModel< M extends AbstractModel< M > > implements M
 			if ( m.getDistance() < epsilon ) inliers.add( m );
 		}
 
-		final double ir = inliers.size() / candidates.size();
+		final double ir = ( double )inliers.size() / ( double )candidates.size();
 		setCost( Math.max( 0.0, Math.min( 1.0, 1.0 - ir ) ) );
 
 		return ( inliers.size() >= minNumInliers && ir > minInlierRatio );
